+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 <!-- markdownlint-disable MD024 -->
 <!-- changelog-start -->
 
-## [Unreleased](https://github.com/DonalChilde/bash-task/v0.0.1...refs/heads/develop)
+## [Unreleased](https://github.com/DonalChilde/bash-task/0.0.1...refs/heads/dev)
 <!-- Dont forget to:
     - Update the Unreleased compare version to latest release tag
     - Update compare/_previous_version_tag_
@@ -10,9 +10,40 @@
 <!-- Copy paste release notes below here -->
 <!-- scriv-insert-here -->
 
-## v0.0.0 - 2022-04-13
+## [0.0.1](https://github.com/DonalChilde/bash-task/compare/0.0.0...0.0.1) —  2023-06-15
 
-### Whats Changed in v0.0.0
+Whats Changed in 0.0.1
+
+### Added
+
+- Define a template task file.
+
+  - closes
+    - [#1](https://github.com/DonalChilde/bash-task/issues/1)
+
+- Add init-venv.sh example.
+
+- Add an example of formatting python code.
+
+  - Pull requests
+    - [#5](https://github.com/DonalChilde/bash-task/pull/5)
+
+  - closes
+    - [#3](https://github.com/DonalChilde/bash-task/issues/3)
+
+### Changed
+
+- Move the python dev dependencies to a requirements.txt file, as `pip install -e .` is not appropriate for this project.
+
+  - Pull requests
+    - [#4](https://github.com/DonalChilde/bash-task/pull/4)
+
+  - closes
+    - [#2](https://github.com/DonalChilde/bash-task/issues/2)
+
+## 0.0.0 - 2023-06-11
+
+### Whats Changed in 0.0.0
 
 This is the start of something....